requirements.txt
Flask==3.0.2
Flask-Cors==4.0.0
python-dotenv==1.0.1
pandas==2.2.1
Werkzeug==3.0.1
gunicorn==21.2.0
pytest==8.0.2 